==> S6_Analysis.qmd <==
---
title: "Inter-rater and inter-method concordance of malarial retinopathy grading"
author: 
  - "Kyle J Wilson"
  - "Alice Muiruri Liomba"
  - "Karl B Seydel"
  - "Owen K Banda"
  - "Christopher A Moxon"
  - "Ian JC MacCormick"
  - "Simon P Harding"
  - "Nicholas AV Beare"
  - "Terrie E Taylor"
format:
  #html: default
  pdf: default
  #docx: default
---

## 1 Introduction 

Indirect ophthalmoscopy is the gold standard test for malarial retinopathy. However, it is impractical to blind an ophthalmologist or clinician from coma status and parasitaemia in a high acuity clinical setting. This could potentially be a source of unmeasured bias.

To demonstrate that grading is consistent, two assessors graded 80 retinal photos of 80 eyes from 41 participants over the same study period. One assessor (KJW) is an ophthalmologist experienced in the assessment of malarial retinopathy. The second assessor (OKB) is an optometrist more recently trained in the assessment of malarial retinopathy. By including graders with differing levels of experience we also demonstrate that agreement is robust to level of training. 

## 2 Analysis
### 2.1 Inter-rater concordance from retinal images

We first show agreement for malarial retinopathy status (positive or negative) and then for the individual signs of malarial retinopathy. First, we prepare the data.

```{r message=FALSE, warning=FALSE}
#| code-fold: true
#| code-summary: "Expand to see the code for data cleaning"

# install and load packages necessary for analysis and displaying the data
packages <- c("readr", "tidyr", "readxl", "stringr", "dplyr", "here", "vcd")

for (pkg in packages) {
  if (!requireNamespace(pkg, quietly = TRUE)) {
    install.packages(pkg)
  }
  library(pkg, character.only = TRUE)
}

set.seed(0)

dual <- read_xlsx(here("../data/ImageConcordanceData.xlsx"))
dual <- dual %>% mutate(across(2:11, ~na_if(.x, -9)))

dual <- dual %>%
  rowwise() %>%
  mutate(status_ob = case_when(
    all(c_across(2:6) == 0, na.rm = FALSE) ~ 0,
    all(is.na(c_across(2:6)) | c_across(2:6) == 0) ~ NA_real_,
    any(c_across(2:6) != 0, na.rm = TRUE) ~ 1),
    status_kw = case_when(
    all(c_across(7:11) == 0, na.rm = FALSE) ~ 0,
    all(is.na(c_across(7:11)) | c_across(7:11) == 0) ~ NA_real_,
    any(c_across(7:11) != 0, na.rm = TRUE) ~ 1)
  ) %>%
  ungroup()
```

Next, we compare grading of the presence or absence of MR.

```{r}
tab <- table(dual$status_ob, dual$status_kw)
res.k <- vcd::Kappa(tab)
ci.k <- confint(res.k)
cat("MR status: ", round(res.k$Unweighted["value"], 3), 
    " (", round(ci.k[2,1], 3), " - ", round(ci.k[2,2], 3), ")\n\n", sep = "")
```

Then we look at individual features of MR.

```{r}
features <- c("ahaem", "acfa", "awhitcmac", "apw", "avc")
feature_names <- c(
  ahaem = "Haemorrhages",
  acfa = "Foveal whitening",
  awhitcmac = "Macular whitening",
  apw = "Peripheral whitening",
  avc = "Vessel change"
)

for (i in seq_along(features)) {
  feat <- features[i]
  name <- feature_names[feat]
  tab <- table(dual[[paste0(feat, "_ob")]], dual[[paste0(feat, "_kw")]])
  res.k <- vcd::Kappa(tab, "Fleiss-Cohen")
  ci.k <- confint(res.k)
  cat(name, ": ", 
      round(res.k$Weighted["value"], 3), 
      " (", round(ci.k[2,1], 3), " - ", round(ci.k[2,2], 3), 
      ")\n\n", sep = "")
}
```

There is substantial agreement (according to Landis and Koch, 1977) in assessment of retinopathy status. There is near perfect agreement in assessment of haemorrhages between the two graders from photos. There is substantial agreement of foveal whitening and macular whitening. There is moderate agreement in both peripheral whitening and vessel change.

These results corroborate findings from Beare et al, 2002 that agreement is good between graders and extend these findings to demonstrate good agreement between graders with different levels of experience and training. 

## 2.2 Inter-method reliability between indirect ophthalmoscopy and image grading

Next, we wanted to compare the data from the photo grading with the findings from clinical examination. First, we prepared the data.

```{r message=FALSE, warning=FALSE}
#| code-fold: true
#| code-summary: "Expand to see the code for data cleaning"

rtd <- read.csv(here("../data/RetinopathyConcordanceData.csv"))

# make cases consistent and remove whitespace
rtd$mpid <- toupper(rtd$mpid)
rtd$mpid <- gsub(" ", "", rtd$mpid)

rtd <- rtd %>% select(1:3,6:7,12:15,24:27,92:95,96,98)
colnames(rtd) <- c("mp_number", "haem_r", "haem_l", "pap_r", "pap_l",
                   "mac_wh_r", "mac_wh_l", "fov_wh_r", "fov_wh_l",
                   "per_wh_r", "per_wh_l", "wh_qu_seen_r", "wh_qu_seen_l",
                   "quad_vc_r", "quad_vc_l", "vc_qu_seen_r", "vc_qu_seen_l",
                   "score", "doc")

# make the data match the photo grading data structure
rtd <- rtd %>% transmute(
  mp_number = mp_number,
  ahaem_r = case_when(
    haem_r == 7 ~ 4,
    haem_r < 0 ~ NA_real_,
    TRUE ~ haem_r), 
  ahaem_l = case_when(
    haem_l == 7 ~ 4,
    haem_l < 0 ~ NA_real_,
    TRUE ~ haem_l),
  acfa_r = case_when(
    fov_wh_r < 0 ~ NA_real_,
    TRUE ~ fov_wh_r),
  acfa_l = case_when(
    fov_wh_l < 0 ~ NA_real_,
    TRUE ~ fov_wh_l),
  awhitcmac_r = case_when(
    mac_wh_r == 3 ~ 2,
    mac_wh_r == 4 ~ 3,
    mac_wh_r < 0 ~ NA_real_,
    TRUE ~ mac_wh_r),
  awhitcmac_l = case_when(
    mac_wh_l == 3 ~ 2,
    mac_wh_l == 4 ~ 3,
    mac_wh_l < 0 ~ NA_real_,
    TRUE ~ mac_wh_l),
  apw_r = if_else(per_wh_r %in% 1:12, 1,
                  if_else(per_wh_r == 0 & 
                         (wh_qu_seen_r == 3 | wh_qu_seen_r == 4), 0, NA_real_)),
  apw_l = if_else(per_wh_l %in% 1:12, 1,
                  if_else(per_wh_l == 0 & 
                         (wh_qu_seen_l == 3 | wh_qu_seen_l == 4), 0, NA_real_)),
  avc_r = if_else(quad_vc_r %in% 1:4, 1, 
                  if_else(quad_vc_r == 0 & 
                         (vc_qu_seen_r == 3 | vc_qu_seen_r == 4), 0, NA_real_)),
  avc_l = if_else(quad_vc_l %in% 1:4, 1,
                  if_else(quad_vc_l == 0 & 
                         (vc_qu_seen_l == 3 | vc_qu_seen_l == 4), 0, NA_real_))
  )

rtd <- rtd %>%
  pivot_longer(
    cols = -mp_number,
    names_to = c(".value", "side"),
    names_sep = "_"
  ) %>%
  mutate(mp_number = paste0(mp_number, toupper(side))) %>%
  select(-side)

rtd <- rtd %>% filter(mp_number %in% dual$mp_number)
rtd <- rtd %>%
  rowwise() %>%
  mutate(status = case_when(
    all(c_across(2:6) == 0, na.rm = FALSE) ~ 0,
    all(is.na(c_across(2:6)) | c_across(2:6) == 0) ~ NA_real_,
    any(c_across(2:6) != 0, na.rm = TRUE) ~ 1)
  ) %>%
  ungroup()

rtd <- left_join(rtd, dual, join_by(mp_number == mp_number))
```

Now, we can conduct the analysis. For this analysis we used the ophthalmologist grading (KJW) and compared to graded ophthalmologist assessment. Once again, first we look at presence and absence of MR.

```{r}
tab <- table(rtd$status, rtd$status_kw)
res.k <- vcd::Kappa(tab)
ci.k <- confint(res.k)
cat("MR status: ", round(res.k$Unweighted["value"], 3), 
    " (", round(ci.k[2,1], 3), " - ", round(ci.k[2,2], 3), ")\n\n", sep = "")
```

Then we look at the individual features of MR. 

```{r}
for (i in seq_along(features)) {
  feat <- features[i]
  name <- feature_names[feat]
  tab <- table(rtd[[feat]], rtd[[paste0(feat, "_kw")]])
  res.k <- vcd::Kappa(tab, "Fleiss-Cohen")
  ci.k <- confint(res.k)
  cat(name, ": ", 
      round(res.k$Weighted["value"], 3), 
      " (", round(ci.k[2,1], 3), " - ", round(ci.k[2,2], 3), 
      ")\n\n", sep = "")
}
```

Here, we can see that while agreement is generally good, agreement in haemorrhage grading is worse than in photos. Peripheral findings also have lower agreement than would be ideal. Because we are interested in retinal haemorrhage status, we next interrogated whether the differences tend to go in one direction or the other.

To do this, we simply subtracted the score from the examination from the grading score from stored images, then calculated the mean. If there was a directional effect this would result in a negative number (undergrading from the photos) or positive number (overgrading from the photos). The significance of this directional effect was assessed using the Mann-Whitney-U test for non-parametric data.

```{r warning=FALSE}
p_values <- numeric(length(features))
mean_diffs <- numeric(length(features))

for (i in seq_along(features)) {
  feat <- features[i]
  photo_col <- paste0(feat, "_kw")
  exam_col <- feat
  
  test <- wilcox.test(rtd[[photo_col]], rtd[[exam_col]], paired = TRUE)
  
  mean_diffs[i] <- mean(rtd[[photo_col]] - rtd[[exam_col]], na.rm = TRUE)
  p_values[i] <- test$p.value
}

adjusted_p <- p.adjust(p_values, method = "holm")

for (i in seq_along(features)) {
  feat <- features[i]
  label <- feature_names[[feat]]
  mean_diff <- mean_diffs[i]
  adj_p <- adjusted_p[i]
  
  grade <- if (mean_diff < 0) "undergraded" else "overgraded"
  miss <- if (mean_diff < 0) "missed" else "recognised erroneously"
  
  direction <- if (adj_p < 0.05) "This is statistically significant." else "This is not statistically significant."
  
  if(feat %in% c("ahaem", "acfa", "awhitcmac")) {
    cat(
    "Feature: ", label, "\n",
    "Mean difference: ", round(mean_diff, 3), "\n",
    "Adjusted p-value: ", signif(adj_p, 3), "\n",
    "Interpretation: Approximately 1 in ", abs(round(1 / mean_diff, 0)), 
    " eyes are ", grade, " by 1 level. \n", direction, "\n\n", sep = ""
    )
  } else {
    cat(
    "Feature: ", label, "\n",
    "Mean difference: ", round(mean_diff, 3), "\n",
    "Adjusted p-value: ", signif(adj_p, 3), "\n",
    "Interpretation: ", label, " is ", miss, " in approximately 1 in ", 
    abs(round(1 / mean_diff, 0)), " eyes. \n", direction, "\n\n", sep = ""
    )
  }
  
}
```

## 3 Conclusions

Taken together, these findings show that, while there is considerable agreement between graders (including graders of different training levels), malarial retinopathy grading by photos alone risks systematically undergrading haemorrhages and vessel change (a peripheral finding).
